# swagger.yaml
openapi: 3.0.3
info:
  title: Anemone Invoice Storage API 3.2
  description: |-
    API description for Team Anemone's Invoice Storage.
    All error strings are placeholders for more detailed messages.
    We have split our API to v1 and v2, check all the endpoints for more details
  version: 1.0.11
...
servers:
  - url: https://anemone1-54805996fbe6.herokuapp.com/
...
components:
  schemas:
    invoiceUploadRequestBody:
      type: object
      properties:
        invoice:
          type: string
          example: invoice.xml
...
    success_true:
      type: boolean
      example: true
...
    error:
      type: object
      properties:
        success:
          $ref: '#/components/schemas/success_false'
        error:
          type: string
          example: error message
...
    invoice:
      type: object
      properties:
        invoiceId:
          $ref: '#/components/schemas/invoiceId'
        invoiceName:
          $ref: '#/components/schemas/invoiceName'
        amount:
          $ref: '#/components/schemas/amount'
        date:
          $ref: '#/components/schemas/date'
        trashed:
          $ref: '#/components/schemas/success_false'
...
    invoiceFetchReturnObject:
      type: object
      properties:
        success:
          $ref: '#/components/schemas/success_true'
        invoice:
          $ref: '#/components/schemas/invoice'
    listReturnObject:
      type: object
      properties:
        success:
          $ref: '#/components/schemas/success_true'
        invoices:
          type: array
          items:
            $ref: '#/components/schemas/invoice'
...
    createUserRequestBody:
      type: object
      properties:
        username:
          $ref: '#/components/schemas/username'
        email:
          $ref: '#/components/schemas/email'
        password:
          $ref: '#/components/schemas/password'
...
    invoiceModifyRequestBody:
      type: object
      properties:
        newInvoiceName:
          $ref: '#/components/schemas/newInvoiceName'
        newAmount:
          $ref: '#/components/schemas/newAmount'
        newDate:
          $ref: '#/components/schemas/newDate'
    okReturnObject:
      type: object
      properties:
        success:
          $ref: '#/components/schemas/success_true'
...
paths:
  /invoices:
    post:
      tags:
        - invoices v1
      summary: Upload a new invoice
      description: Upload a new invoice, returning a server-generated invoiceId
      parameters:
        - name: token
          in: header
          description: The token used for user login
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/invoiceUploadRequestBody'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/invoiceFetchReturnObject'
        '401':
          description: Token is empty or invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
    get:
      tags:
        - invoices v1
      summary: Get a list of all invoices
      description: Lists all existing invoices for the currently logged in user that are not trashed
      parameters:
        - name: token
          in: header
          description: The token used for user login
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/listReturnObject'
        '401':
          description: Token is empty or invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
  /invoices/{invoiceId}:
    get:
      tags:
        - invoices v1
      summary: Retrieve an invoice
      description: Retrive an invoice based on a provided invoiceId
      parameters:
        - name: token
          in: header
          description: The token used for user login
          required: true
          schema:
            type: string
        - name: invoiceId
          in: path
          description: The specified invoiceId to retrieve
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/invoice'
        '400':
          description: |-
            If any of the following is true:
            - invoiceId does not refer to an existing invoice
            - invoiceId refers to an invoice in the trash
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        '401':
          description: Token is empty or invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        '403':
          description: Valid token and invoiceId are provided, but user is not owner of this invoice
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
    put:
      tags:
        - invoices v1
      summary: Modify an existing invoice
      description: Allows the user to modify an invoice that has already been stored. If any of the fields are left empty, this will not edit the existing information of that field. For example, an empty string for newInvoiceName would mean that the existing invoiceName is not edited. Any field may be left empty to not edit it, except for all fields being empty which is an error case.
      parameters:
        - name: token
          in: header
          description: The token used for user login
          required: true
          schema:
            type: string
        - name: invoiceId
          in: path
          description: The specified invoiceId to modify
          required: true
          schema:
            type: integer
      requestBody:
        description: Modifying name
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/invoiceModifyRequestBody'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/invoice'
        '400':
          description: |-
            If any of the following is true:
            - invoiceId does not refer to an existing invoice
            - invoiceId refers to an invoice in the trash
            - newAmount is not a positive number and not an empty string ""
            - newDate is an invalid date and not an empty string ""
            - newDate refers to a time in the future
            - all 3 newInvoiceName, newAmount and newDate are empty strings ""
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        '401':
          description: Token is empty or invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        '403':
          description: Valid token and invoiceId are provided, but user is not owner of this invoice
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
    delete:
      tags:
        - trash v1
      summary: Moves an existing invoice into the trash
      description: Moves an existing invoice into the trash
      parameters:
        - name: token
          in: header
          description: The token used for user login
          required: true
          schema:
            type: string
        - name: invoiceId
          in: path
          description: The specified invoiceId to retrieve
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/okReturnObject'
        '400':
          description: |-
            If any of the following is true:
            - invoiceId does not refer to an existing invoice
            - invoiceId refers to an invoice in the trash
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        '401':
          description: Token is empty or invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        '403':
          description: Valid token and invoiceId are provided, but user is not owner of this invoice
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
  /users:
    post:
      tags:
        - users v1
      summary: Create a new user
      description: Creates a new user, with a username, email and password
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/createUserRequestBody'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/createUserReturnObject'
        '400':
          description: |-
            If any of the following is true:
            - username is not between 3-20 characters long
            - username contains empty space " "
            - username is currently used by another user
            - password does not contain at least 1 uppercase letter
            - password does not contain at least 1 lowercase letter
            - password does not contain at least 1 number
            - password does not contain at least 1 special character
            - password does not contain at least 8 characters
            - email does not satisfy this: https://www.npmjs.com/package/validator (validator.isEmail)
            - email is currently used by another user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        '403':
          description: If username was taken by another user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
  /users/login:
    post:
      tags:
        - users v1
      summary: Login a user
      description: Allows a user to login to the system and returns a valid token session given username and password
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/userLoginRequestBody'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/userLoginReturnObject'
        '401':
          description: |-
            If any of the following is true:
            - username does not refer to an existing user
            - password does not match username
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
  /trash/{invoiceId}:
    delete:
      tags:
        - trash v1
      summary: Deletes a trashed invoice
      description: Deletes an invoice completely
      parameters:
        - name: token
          in: header
          description: The token used for user login
          required: true
          schema:
            type: string
        - name: invoiceId
          in: path
          description: The specified invoiceId to retrieve
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/okReturnObject'
        '400':
          description: |-
            If any of the following is true:
            - invoiceId does not refer to an existing invoice
            - invoiceId refers to an invoice not in the trash
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        '401':
          description: Token is empty or invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        '403':
          description: Valid token and invoiceId are provided, but user is not owner of this invoice
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
  /trash/{invoiceId}/restore:
    post:
      tags:
        - trash v1
      summary: Restores a trashed invoice
      description: Moves an invoice from the trash back to normal storage
      parameters:
        - name: token
          in: header
          description: The token used for user login
          required: true
          schema:
            type: string
        - name: invoiceId
          in: path
          description: The specified invoiceId to retrieve
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/okReturnObject'
        '400':
          description: |-
            If any of the following is true:
            - invoiceId does not refer to an existing invoice
            - invoiceId refers to an invoice not in the trash
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        '401':
          description: Token is empty or invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        '403':
          description: Valid token and invoiceId are provided, but user is not owner of this invoice
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
  /trash:
    get:
      tags:
        - trash v1
      summary: Lists trash items
      description: Lists all the invoices that are currently in the trash, in the order that each invoice was deleted; so the latest deleted invoice will be at the bottom of the returned invoices array.
      parameters:
        - name: token
          in: header
          description: The token used for user login
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/trashListReturnObject'
        '401':
          description: Token is empty or invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
...
  /invoicesV2:
    post:
      tags:
        - invoices v2
      summary: Upload a new invoice
      description: Upload a new invoice, returning a server-generated invoiceId
      parameters:
        - name: token
          in: header
          description: The token used for user login
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/invoiceUploadRequestBodyV2'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/invoiceUploadReturnObjectV2'
        '400':
          description: Invoice is not XML format
          content:
              application/json:
                schema:
                  $ref: '#/components/schemas/error'
        '401':
          description: Token is empty or invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
    get:
      tags:
        - invoices v2
      summary: Get a list of all invoices
      description: Lists all existing invoices for the currently logged in user that are not trashed
      parameters:
        - name: token
          in: header
          description: The token used for user login
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/listReturnObjectV2'
        '401':
          description: Token is empty or invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
  /invoicesV2/{invoiceId}:
    get:
      tags:
        - invoices v2
      summary: Retrieve an invoice
      description: Retrive an invoice based on a provided invoiceId
      parameters:
        - name: token
          in: header
          description: The token used for user login
          required: true
          schema:
            type: string
        - name: invoiceId
          in: path
          description: The specified invoiceId to retrieve
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/invoiceRetrieveReturnObjectV2'
        '400':
          description: |-
            If any of the following is true:
            - invoiceId does not refer to an existing invoice
            - invoiceId refers to an invoice in the trash
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        '401':
          description: Token is empty or invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        '403':
          description: Valid token and invoiceId are provided, but user is not owner of this invoice
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
    put:
      tags:
        - invoices v2
      summary: Modify an existing invoice
      description: Allows the user to modify an invoice that has already been stored. If any of the fields are left empty, this will not edit the existing information of that field. For example, an empty string for newInvoiceName would mean that the existing invoiceName is not edited. Any field may be left empty to not edit it, except for all fields being empty which is an error case.
      parameters:
        - name: token
          in: header
          description: The token used for user login
          required: true
          schema:
            type: string
        - name: invoiceId
          in: path
          description: The specified invoiceId to modify
          required: true
          schema:
            type: integer
      requestBody:
        description: Modifying name
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/invoiceModifyRequestBody'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/invoiceRetrieveReturnObjectV2'
        '400':
          description: |-
            If any of the following is true:
            - invoiceId does not refer to an existing invoice
            - invoiceId refers to an invoice in the trash
            - newAmount is not a positive number and not an empty string ""
            - newDate is an invalid date and not an empty string ""
            - newDate refers to a time in the future
            - all 3 newInvoiceName, newAmount and newDate are empty strings ""
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        '401':
          description: Token is empty or invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        '403':
          description: Valid token and invoiceId are provided, but user is not owner of this invoice
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
    delete:
      tags:
        - trash v2
      summary: Moves an existing invoice into the trash
      description: Moves an existing invoice into the trash
      parameters:
        - name: token
          in: header
          description: The token used for user login
          required: true
          schema:
            type: string
        - name: invoiceId
          in: path
          description: The specified invoiceId to retrieve
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/okReturnObject'
        '400':
          description: |-
            If any of the following is true:
            - invoiceId does not refer to an existing invoice
            - invoiceId refers to an invoice in the trash
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        '401':
          description: Token is empty or invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        '403':
          description: Valid token and invoiceId are provided, but user is not owner of this invoice
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
  /usersV2:
    post:
      tags:
        - users v2
      summary: Create a new user
      description: Creates a new user, with a username, email and password
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/createUserRequestBody'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/createUserReturnObject'
        '400':
          description: |-
            If any of the following is true:
            - username is not between 3-20 characters long
            - username contains empty space " "
            - username is currently used by another user
            - password does not contain at least 1 uppercase letter
            - password does not contain at least 1 lowercase letter
            - password does not contain at least 1 number
            - password does not contain at least 1 special character
            - password does not contain at least 8 characters
            - email does not satisfy this: https://www.npmjs.com/package/validator (validator.isEmail)
            - email is currently used by another user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        '403':
          description: If username was taken by another user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
  /usersV2/login:
    post:
      tags:
        - users v2
      summary: Login a user
      description: Allows a user to login to the system and returns a valid token session given username and password
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/userLoginRequestBody'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/userLoginReturnObject'
        '401':
          description: |-
            If any of the following is true:
            - username does not refer to an existing user
            - password does not match username
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
...
  /trashV2:
    get:
      tags:
        - trash v2
      summary: Lists trash items
      description: Lists all the invoices that are currently in the trash, in the order that each invoice was deleted; so the latest deleted invoice will be at the bottom of the returned invoices array.
      parameters:
        - name: token
          in: header
          description: The token used for user login
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/listReturnObjectV2'
        '401':
          description: Token is empty or invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
...